refactor background task into separate module
