+++ busy_beaver/backend/tasks.py
@@ -1,12 +1,8 @@
-from datetime import timedelta
 import logging
-from typing import List
 
-from sqlalchemy import and_
-
-from .. import api, db, github_stats, slack
-from ..models import ApiUser, User
-from ..toolbox import utc_now_minus
+from .. import db
+from ..models import ApiUser
+from ..tasks import post_github_summary_to_slack
 
 logger = logging.getLogger(__name__)
 
@@ -48,19 +44,3 @@ class PublishGitHubSummaryResource:
 
         logger.info("[Busy-Beaver] Post GitHub Summary -- kicked-off")
         resp.media = {"run": "kicked_off"}
-
-
-@api.background.task
-def post_github_summary_to_slack(channel: str) -> None:
-    boundary_dt = utc_now_minus(timedelta(days=1))
-    channel_id = slack.get_channel_id(channel)
-    members = slack.get_channel_members(channel_id)
-
-    users: List[User] = db.query(User).filter(
-        and_(User.slack_id.in_(members), User.github_username.isnot(None))
-    ).all()
-    message = ""
-    for user in users:
-        message += github_stats.generate_summary(user, boundary_dt)
-
-    slack.post_message(channel_id, message)
